Improved CMake clean script

+++ tools/cmakeclean.sh
@@ -1,7 +1,13 @@
 #!/usr/bin/env bash
 
-rm CMakeCache.txt
-rm -rf CMakeFiles
-rm cmake_install.cmake
-rm Makefile
-rm -rf src
+pwd=${PWD##*/}
+
+if [[ $pwd == "build" ]]; then
+    rm CMakeCache.txt
+    rm -rf CMakeFiles
+    rm cmake_install.cmake
+    rm Makefile
+    rm -rf src
+else
+    echo "ERROR: This script must be run from the 'build' directory."
+fi
